+++ wren-ai-service/src/core/provider.py
@@ -3,6 +3,20 @@ from abc import ABCMeta, abstractmethod
 from haystack.document_stores.types import DocumentStore
 
 
+class LLMProvider(metaclass=ABCMeta):
+    @abstractmethod
+    def get_generator(self, *args, **kwargs):
+        ...
+
+    @abstractmethod
+    def get_text_embedder(self, *args, **kwargs):
+        ...
+
+    @abstractmethod
+    def get_document_embedder(self, *args, **kwargs):
+        ...
+
+
 class DocumentStoreProvider(metaclass=ABCMeta):
     @abstractmethod
     def get_store(self, *args, **kwargs) -> DocumentStore:
