+++ wren-ai-service/src/pipelines/ask/historical_question.py
@@ -4,9 +4,8 @@ from typing import Dict, List, Optional
 
 from haystack import Document, Pipeline, component
 
-from src.core.document_store_provider import DocumentStoreProvider
-from src.core.llm_provider import LLMProvider
 from src.core.pipeline import BasicPipeline
+from src.core.provider import DocumentStoreProvider, LLMProvider
 from src.utils import (
     init_providers,
     load_env_vars,
